Project refactoring.
Change project types to use LLVM-clangcl toolchain. I did this because it appears that Microsoft does not support their MSVC compiler much. Enable /Wall /Werror where possible, selectively disable warnings for things I can't fix right now. Do a lot of refactoring to appease new compiler settings. Leave debug with address sanitaizer build on MSVC++ compiler (v145) because clangcl does not properly support it for some reason, even though I am pretty sure that it uses address sanitaizer developed by clang

+++ program_loader/function_call_builder.h
@@ -69,7 +69,7 @@ public:
 		}
 		else {
 			variable_imm<std::uint64_t> immediate_function_displacement{
-				static_cast<std::uint64_t>(this->get_function_table_index(fnc->get_id()))
+				(this->get_function_table_index(fnc->get_id()))
 			};
 
 			this->place_immediate(&immediate_function_displacement);
